+++ cypress-shared/integration/b2b/2.4.3-show_organization_by_adding_principal_sales_channel.spec.js
@@ -22,7 +22,7 @@ describe('Add binding which shows Organization in profile page', () => {
 
   verifySalesChannel(1)
 
-  verifyBindings(users.OrganizationAdmin1, true)
+  verifyBindings(users.OrganizationAdmin1.email, true)
 
   loginToStoreFront(
     users.OrganizationAdmin1,