+++ cypress-shared/integration/b2b/2.5.1-buyer_scenarios_orgA_costA1.spec.js
@@ -32,25 +32,24 @@ describe('Organization A - Cost Center A1 - Buyer Scenarios', () => {
     users,
     product,
     quotes,
-    gmailCreds,
   } = b2b.OrganizationA
 
   const { organizationName: organizationB, quotes: organizationBQuote } =
     b2b.OrganizationB
 
-  loginToStoreFront(users.Buyer1, ROLE_DROP_DOWN.Buyer, gmailCreds)
+  loginToStoreFront(users.Buyer1, ROLE_DROP_DOWN.Buyer)
   verifySession(b2b.OrganizationA, costCenter1.name, ROLE_DROP_DOWN.Buyer)
   productShouldNotbeAvailableTestCase(nonAvailableProduct)
   userAndCostCenterShouldNotBeEditable({
     organizationName,
     costCenter: costCenter1.name,
-    gmailCreds,
+    email: users.Buyer1.email,
     role: role.Buyer1,
   })
   userShouldNotImpersonateThisUser(
     ROLE_DROP_DOWN.Buyer,
     roleObject.SalesManager.role,
-    users.SalesManager
+    users.SalesManager.email
   )
   quoteShouldbeVisibleTestCase(
     organizationName,