+++ cypress-shared/integration/b2b/2.7-buyer_scenarios_orgB_costB1.spec.js
@@ -25,13 +25,12 @@ describe('Organization B - Cost Center B1 - Buyer Scenario', () => {
     product,
     quotes,
     costCenter1,
-    gmailCreds,
   } = b2b.OrganizationB
 
   const { organizationName: organizationA, quotes: organizationAQuotes } =
     b2b.OrganizationA
 
-  loginToStoreFront(users.Buyer1, ROLE_DROP_DOWN.Buyer, gmailCreds)
+  loginToStoreFront(users.Buyer1, ROLE_DROP_DOWN.Buyer)
   verifySession(b2b.OrganizationB, costCenter1.name, ROLE_DROP_DOWN.Buyer)
   productShouldNotbeAvailableTestCase(nonAvailableProduct)
   quoteShouldbeVisibleTestCase(