+++ CHANGELOG.md
@@ -18,8 +18,13 @@
 
 
 ### Changed
+- Improved/Fixed B2B new login user testcase
+
+### Changed
 
 - E2E Tests improvements
+
+- [ENGINEERS-1228] - Use multiple email addresses in B2B tests
 
 ## [2.4.2] - 2023-03-12